Fix for focus and target
This has taken me much more time than I'd like. But for some reason the on click function receives different parameters when not overwritten in sArena.xml resulting in typerelease instead of type. type is correctly returned when overwriting onclick in sarena.xml however this is not allowed as this is a secure function and cannot be called. As I am not sure what is the difference or how to steer it properly for now I am leaving type1 and type2 attributes as well as adding typerelease ones.