Fixes for 10.0 (#66)
* Fixes for 10.0

Changes required by new handling of cast bar. Kept back compatiblity with color setting as much as I could. best to my knowledge

* Fix to hide Blizzard frames

Fix to hide Blizzard frames

* Fix to hiding game frames

* Valida frame names

* Updated version of Ace

It is weird as previous commits did not pop issues with Ace, even when sarena was run alone. However now it has so I updated the Ace libraries.

* Fix for focus and target

This has taken me much more time than I'd like. But for some reason the on click function receives different parameters when not overwritten in sArena.xml resulting in typerelease instead of type. type is correctly returned when overwriting onclick in sarena.xml however this is not allowed as this is a secure function and cannot be called. As I am not sure what is the difference or how to steer it properly for now I am leaving type1 and type2 attributes as well as adding typerelease ones.

* Fix for focus and target when not using press and cast and ActionButtonUseKeyDown is true

From version 10 games checkes, before perfoming registered action, whether user has ActionButtonUseKeyDown set to true or false. If it does not match registered click, the action is rejected. Therfore it is needed to handle both situation so regardless of user setting the registered actions will work.

* Removed tabs

* Updated version to 3.2.0

* Removed tabs from xml file

* Removed commented out line

* Removed previous check of Blizzard Arena addon loaded

This addon no longer exists and seems it was changed to normal in game frame.

* Additional formating fixes

* Reverting incorrectly setup parameter for debugging

Not sure how that got commited in, reverting this change (used during debugging)

* Tab removals and format fixes in layouts

* Added new globals to pass the lua check

* Changed order